Implement dir stripping
Source archive often includes leading directories. Allow stripping.

Package `tar` supports leading directory stripping. Package
`extract-zip` does not. Implement a library to perform path stripping in
similar behavior so we can bridge the gap.

While we're here, expose stripping logic in the extraction and the
download internal libs. For tar, this is passed on to package `tar`. For
zip, implement the stripping behavior to mimic that of tar.