feat: Customizable target directory layout
Layout of the source directory may not necessarily be the desired layout
for the target directory. Move the building process to a temporary work
directory. Next, implement a way for consumers to customize the target
directory layout (by way of remapping source => target). Lastly, move
the contents of the build directory to the target directory, falling
back to moving all contents if no remappings are specified.